👌 REFACTOR: Related to Centralized Interfaces And Types
