+++ src/public-api.ts
@@ -6,5 +6,5 @@ export * from './lib/hotkeys.service';
 export * from './lib/hotkeys.directive';
 export * from './lib/hotkeys-cheatsheet/hotkeys-cheatsheet.component';
 export * from './lib/hotkey.model';
-export * from './lib/hotkey.options';
+export * from './lib/hotkey.interfaces-types';
 export * from './lib/hotkey.module';
